+++ src/index.ts
@@ -1,9 +1,11 @@
 export { OphButton } from './components/OphButton';
 export { OphCheckbox } from './components/OphCheckbox';
-export { OphSelect, OphSelectFormField } from './components/OphSelect';
+export { OphSelect } from './components/OphSelect';
+export { OphSelectFormField } from './components/OphSelectFormField';
 export { OphTypography } from './components/OphTypography';
 export { OphLink } from './components/OphLink';
-export { OphInput, OphInputFormField } from './components/OphInput';
+export { OphInput } from './components/OphInput';
+export { OphInputFormField } from './components/OphInputFormField';
 export { ophColors } from './ophColors';
 export { OphFormFieldWrapper } from './components/OphFormFieldWrapper';
