+++ src/story-common.ts
@@ -1,0 +1,4 @@
+export const OPTIONS = [
+  { value: '1', label: 'First option' },
+  { value: '2', label: 'Second option' },
+];
